Handle long Java classpaths.
The groovyc command requires that the Java classpath be specified as one
argument, which does not work for long classpaths, since Linux limits the
length of any one command argument to 131072 characters.  Work around the
problem by creating symlinks with short names to classpath jars and using the
symlinks instead of the original jar file name paths.